requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
google-genai>=0.3.0
openai>=1.3.0
plotly>=5.15.0
sqlparse>=0.4.0
networkx>=3.0.0
python-dotenv>=1.0.0
antlr4-python3-runtime>=4.13.0
